+++ lib/srv/desktop/rdp/rdpclient/src/rdpdr/tdp.rs
@@ -589,7 +589,7 @@ pub struct SharedDirectoryTruncateRequest {
     pub completion_id: u32,
     pub directory_id: u32,
     pub path: UnixPath,
-    pub end_of_file: u32,
+    pub end_of_file: i64,
 }
 
 impl SharedDirectoryTruncateRequest {
